Return holidays for provinces
Earlier, we had been ignoring holidays in the output for provinces,
but that meant we couldn't easily see all holidays for a given
province (except through query parameters).

Removing the @JsonIgnore annotation created an infininininite
recursion situation, so that wasn't going to work.

Luckily, the @JsonIgnoreProperties annotation lets us ignore
nested fields, which is perfect.

Now:
- when we search for provinces, we also see their holidays
- when we search for holidays, we also see their provinces

But nested holidays don't have provinces, while nested provinces
don't have holidays, if that makes sense.

This was the link that led me to the answer:
- http://springquay.blogspot.com/2016/01/new-approach-to-solve-json-recursive.html