Run the app on port 80 rather than port 8080
Running this as a docker container, we can map port 80 to the
application's default port 8080 with the command

`docker run -p 80:8080 holidays`

Unfortunately, this doesn't get picked up on Azure's Container
Instances service, so it looks like the fix is to run the application
on port 80.

Also updated the container name to match the Dockerhub repo.