Hardcode date strings with current year
The way the APi works, it takes a human readable date string
from Wikipedia and then it interprets it into a Java Date.

So with a string like "Third Wednesday of April", it will go ahead
and convert that into a useable ISO date.

However, since there's no year information, it will just take a guess.
In our case, if the date had already passed for this year, it finds
the date for next year.

Technically it works as expected, but since it's not what we want
our API to do, it's a bug. Hardcoding the year to the end of the
date string works for 100% of our holiday strings so that's probably
fine until we need to make this more complicated.